Initial Commit
Added Processing project for the first time. I messed up some values
somewhere and the thumbDetection is a bit less accurate than it was an
hour before, but I know I can fix that. The finger detection algorithm
is literally taken from the previous version, and therefore isn't used
to some of the new values. That too can be altered.